Add docs and specs, refactor functions and add bang versions,
also update README and bump version

This commit mostly makes the user functions act in a more idiomatic way,
returning {:ok, value} and {:error, reason} tuples unless the bang versions
of the functions are used. It also adds specs and documentation for every
function, as well as updating the README to tell how to generate docs.